# Improve Hausdorff perf and accept larger number of inputs.