--- a/cpp/src/spatial/hausdorff.cu
+++ b/cpp/src/spatial/hausdorff.cu
@@ -142,7 +142,6 @@ struct hausdorff_functor {
     stream.synchronize();
 
     for (uint32_t i = 1; i <= h_space_offsets.size(); i++) {
-
       uint32_t elements_in_batch = 0;
 
       auto starting_i = i;
@@ -150,16 +149,13 @@ struct hausdorff_functor {
       // deduce appropriate batch size based on input.
 
       while (i <= h_space_offsets.size()) {
-
-        uint32_t space_size = (i < h_space_offsets.size() ? h_space_offsets[i] : xs.size()) - h_space_offsets[i - 1];
+        uint32_t space_size =
+          (i < h_space_offsets.size() ? h_space_offsets[i] : xs.size()) - h_space_offsets[i - 1];
         uint32_t next_size = xs.size() * space_size;
-        if (elements_in_batch > std::numeric_limits<int32_t>::max() - next_size) {
-          break;
-        }
+        if (elements_in_batch > std::numeric_limits<int32_t>::max() - next_size) { break; }
         elements_in_batch += next_size;
         i++;
       }
-
 
       thrust::inclusive_scan_by_key(rmm::exec_policy(stream),
                                     gcp_key_iter,