--- a/cpp/src/spatial/hausdorff.cu
+++ b/cpp/src/spatial/hausdorff.cu
@@ -80,40 +80,53 @@ __global__ void kernel_hausdorff(size_type num_points,
                                  size_type const* space_offsets,
                                  T* results)
 {
+  // determine the LHS point this thread is responsible for.
   auto const thread_idx = blockIdx.x * blockDim.x + threadIdx.x;
   auto const lhs_p_idx  = thread_idx;
 
   if (lhs_p_idx >= num_points) { return; }
 
+  // detemine the LHS space this point belongs to.
   auto const lhs_space_idx =
     thrust::distance(
       space_offsets,
       thrust::upper_bound(thrust::seq, space_offsets, space_offsets + num_spaces, lhs_p_idx)) -
     1;
 
-  // get the point this thread will be computing against for the lhs.
+  // get the x and y coordinate of this LHS point.
   auto const lhs_p_x = xs[lhs_p_idx];
   auto const lhs_p_y = ys[lhs_p_idx];
 
+  // loop over each RHS space, as determined by space_offsets
   for (uint32_t rhs_space_idx = 0; rhs_space_idx < num_spaces; rhs_space_idx++) {
+    
+    // determine the begin/end offsets of points contained within this RHS space.
     auto const rhs_p_idx_begin = space_offsets[rhs_space_idx];
     auto const rhs_p_idx_end =
       (rhs_space_idx + 1 == num_spaces) ? num_points : space_offsets[rhs_space_idx + 1];
 
+    // each space must contain at least one point, this initial value is just an identity value to
+    // simplify calculations.
     auto min_distance_squared = std::numeric_limits<T>::max();
 
+    // loop over each point in the current RHS space
     for (uint32_t rhs_p_idx = rhs_p_idx_begin; rhs_p_idx < rhs_p_idx_end; rhs_p_idx++) {
+
+      // get the x and y coordinate of this RHS point
       auto const rhs_p_x = xs[rhs_p_idx];
       auto const rhs_p_y = ys[rhs_p_idx];
 
-      // get distance between lhs_p and rhs_p
+      // get distance between the LHS and RHS point
       auto const distance_squared = magnitude_squared(rhs_p_x - lhs_p_x, rhs_p_y - lhs_p_y);
 
+      // remember only smallest distance from this LHS point to any RHS point.
       min_distance_squared = min(min_distance_squared, distance_squared);
     }
 
+    // determine the output offset for this pair of spaces (LHS, RHS)
     auto output_idx = lhs_space_idx * num_spaces + rhs_space_idx;
 
+    // use atomicMax to find the maximum of the minimum distance calculated for each space pair.
     atomicMax(results + output_idx, sqrt(min_distance_squared));
   }
 }