--- a/cpp/src/spatial/hausdorff.cu
+++ b/cpp/src/spatial/hausdorff.cu
@@ -159,10 +159,7 @@ struct hausdorff_functor {
     auto const result_view = result->mutable_view();
 
     // due to hausdorff kernel using `atomicMax` for output, the output must be initialized to 0.
-    thrust::fill(rmm::exec_policy(stream),
-                 result_view.begin<T>(),
-                 result_view.end<T>(),
-                 0);
+    thrust::fill(rmm::exec_policy(stream), result_view.begin<T>(), result_view.end<T>(), 0);
 
     auto const threads_per_block = 64;
     auto const num_tiles         = (num_points + threads_per_block - 1) / threads_per_block;