@cwharris 2021-06-26
One of the hausdorff tests was failing due to `atomicMax` being used against some uninitialized memory. Initializing the output did not significantly impact performance.
```
------------------------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------
HausdorffBenchmark/hausdorff/32/4/manual_time          0.057 ms        0.215 ms         9811 items_per_second=152.51M/s
HausdorffBenchmark/hausdorff/64/4/manual_time          0.097 ms        0.255 ms         7132 items_per_second=368.497M/s
HausdorffBenchmark/hausdorff/512/4/manual_time         0.825 ms        0.982 ms          829 items_per_second=2.84929G/s
HausdorffBenchmark/hausdorff/4096/4/manual_time         8.81 ms         10.9 ms           68 items_per_second=17.1311G/s
HausdorffBenchmark/hausdorff/8192/4/manual_time         30.9 ms         33.0 ms           22 items_per_second=19.522G/s
HausdorffBenchmark/hausdorff/32/8/manual_time          0.075 ms        0.234 ms         9317 items_per_second=626.194M/s
HausdorffBenchmark/hausdorff/64/8/manual_time          0.133 ms        0.292 ms         5238 items_per_second=1.45975G/s
HausdorffBenchmark/hausdorff/512/8/manual_time          1.12 ms         1.27 ms          620 items_per_second=11.4586G/s
HausdorffBenchmark/hausdorff/4096/8/manual_time         27.8 ms         29.8 ms           25 items_per_second=29.5844G/s
HausdorffBenchmark/hausdorff/8192/8/manual_time          104 ms          106 ms            7 items_per_second=31.7485G/s
HausdorffBenchmark/hausdorff/32/64/manual_time         0.273 ms        0.432 ms         2555 items_per_second=13.9497G/s
HausdorffBenchmark/hausdorff/64/64/manual_time         0.536 ms        0.694 ms         1261 items_per_second=29.4053G/s
HausdorffBenchmark/hausdorff/512/64/manual_time         23.2 ms         23.3 ms           30 items_per_second=44.7398G/s
HausdorffBenchmark/hausdorff/4096/64/manual_time        1459 ms         1459 ms            1 items_per_second=45.6102G/s
HausdorffBenchmark/hausdorff/8192/64/manual_time        5917 ms         5917 ms            1 items_per_second=45.0033G/s
HausdorffBenchmark/hausdorff/32/128/manual_time        0.503 ms        0.660 ms         1372 items_per_second=30.8398G/s
HausdorffBenchmark/hausdorff/64/128/manual_time         1.65 ms         1.81 ms          421 items_per_second=38.8039G/s
HausdorffBenchmark/hausdorff/512/128/manual_time        96.7 ms         96.9 ms            7 items_per_second=43.5433G/s
HausdorffBenchmark/hausdorff/4096/128/manual_time       5849 ms         5849 ms            1 items_per_second=46.2435G/s
HausdorffBenchmark/hausdorff/8192/128/manual_time      23490 ms        23490 ms            1 items_per_second=46.0679G/s
```